+++ HISTORY.md
@@ -7,6 +7,7 @@
 #### API
 
 - Add `GradientAborescenceEmitter`, which is used to implement CMA-MEGA (#240)
+- Specify that `x0` is 1D for all emitters (#244)
 - Add `best_elite` property for archives (#237)
 - Rename methods in ArchiveDataFrame and rename as_pandas behavior columns
   (#236)